restructure, moving the mid-level game objects
